requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
pytest
httpx
python-dotenv
requests
sentry-sdk

